change the type of readTimeout from int to long

+++ httpclient-core/src/main/java/esa/httpclient/core/CompositeRequest.java
@@ -276,7 +276,7 @@ public class CompositeRequest extends HttpRequestBaseImpl implements PlainReques
     }
 
     @Override
-    public CompositeRequest readTimeout(int readTimeout) {
+    public CompositeRequest readTimeout(long readTimeout) {
         checkStarted();
         super.readTimeout(readTimeout);
         return self();
