+++ httpclient-core/src/main/java/esa/httpclient/core/Request.java
@@ -56,5 +56,5 @@ public interface Request {
      *
      * @return readTimeout
      */
-    int readTimeout();
+    long readTimeout();
 }
